use explicit casts

+++ Extractor/XMLParser.cpp
@@ -279,13 +279,13 @@ ImportNode XMLParser::ReadXMLNode()
     xmlChar *attribute = xmlTextReaderGetAttribute(inputReader, (const xmlChar *)"lat");
     if (attribute != NULL)
     {
-        node.lat = COORDINATE_PRECISION * StringToDouble((const char *)attribute);
+        node.lat = static_cast<int>(COORDINATE_PRECISION * StringToDouble((const char *)attribute));
         xmlFree(attribute);
     }
     attribute = xmlTextReaderGetAttribute(inputReader, (const xmlChar *)"lon");
     if (attribute != NULL)
     {
-        node.lon = COORDINATE_PRECISION * StringToDouble((const char *)attribute);
+        node.lon = static_cast<int>(COORDINATE_PRECISION * StringToDouble((const char *)attribute));
         xmlFree(attribute);
     }
     attribute = xmlTextReaderGetAttribute(inputReader, (const xmlChar *)"id");
